Fix bugs in fabric.invoke and add registerUser
fabric.invoke() was not returning anything on success, so we couldn't put anything in "data".\nfabric.registerUser does not return anything on success because this is only adding the user to wallet. Server then uses invoke to addUser() to the state which returns the user asset that was added to the world state and the server then puts that returned object, i.e. contractResponse in "data" and responds